[frontend] Fix the quick Subscription button
